Add data mapping capabilities
- Add 2 more arguments to `useData` hook: first function is used
to map actual data to a desirable value, and the third one can
be used to generate comparable value based on mapped result. This
functionality can be used to reduce amount of renders: when actual
data changes, the piece of it that is actually used by hook
consumer can actually stay the same.
- Also, add one more `useEffect` hook to mark handler on component
unmount to avoid nasty side-effects